+++ ETS2LA/Program.cs
@@ -1,7 +1,10 @@
 ﻿using Velopack;
 using Velopack.Locators;
 
+using ETS2LA.UI;
 using ETS2LA.Overlay;
+using ETS2LA.Backend;
+using ETS2LA.Telemetry;
 
 namespace ETS2LA;
 
@@ -13,6 +16,8 @@ internal static class Program
     static void Main(string[] args)
     {
         // Velopack is the installer / update manager
+        // Please don't move this, Velopack has to be initialized before anything else,
+        // otherwise we might end up with weird bugs.
         VelopackApp.Build()
             .SetAutoApplyOnStartup(false)
             #if DEBUG
@@ -26,15 +31,27 @@ internal static class Program
 
         var BackendThread = Task.Run(() =>
         {
-            // These initialize global instances of both the overlay and the backend.
-            // Overlay is started "first" since some plugins might need to reference it.
+            // These initialize global instances, if there's a more "official" way to
+            // do this then please make a PR for that.
             var ar = OverlayHandler.Current;
-            var backend = Backend.PluginBackend.Current;
+            var backend = PluginBackend.Current;
+            var telemetry = GameTelemetry.Current;
         });
+
+        # if LINUX
+            string? useWayland = Environment.GetEnvironmentVariable("GLFW_USE_WAYLAND");
+            if (useWayland == null || useWayland == "0" || useWayland == "")
+            {
+                // This is to prevent GLFW from trying to use wayland. If wayland is still required
+                // then setting GLFW_USE_WAYLAND=1 should work fine.
+                Environment.SetEnvironmentVariable("GLFW_USE_WAYLAND", "0");
+                Environment.SetEnvironmentVariable("SDL_VIDEODRIVER", "x11");
+            }
+        # endif
 
         // Gotta wait for the UI thread to close (i.e. user closed the window)
         // and then tell the backend to shutdown too.
         UI.Program.Main(args);
-        Backend.PluginBackend.Current.Shutdown();
+        PluginBackend.Current.Shutdown();
     }
 }