+++ Plugins/VisualizationSockets/Channels/Traffic.cs
@@ -1,11 +1,11 @@
 using ETS2LA.Logging;
 using ETS2LA.Shared;
-using ETS2LA.Backend;
-using System.Net.WebSockets;
+using ETS2LA.Game.SDK;
 using ETS2LA.Backend.Events;
+
+using System.Net.WebSockets;
 using System.Text.Json.Serialization;
 using System.Text.Json;
-using Avalonia.Remote.Protocol.Designer;
 
 namespace VisualizationSockets.Channels;